Make finding the correct widget in the test app more robust
This finds the correct widget regardless of whether Gtk renders its own
decoration: If Gtk renders its own window title bar, the box containing
the label and button is at index 1. If Gtk does not render a title bar,
the box is at index 0. To find the box, look for the frame child with
role :filler. The title bar has role :panel.